update the test for article
